Made a lot of changes, to hopefully make the code cross-platform. There are some memory issues, try to see if those can be cleared up.

+++ src/histogram_gn.h
@@ -3,12 +3,13 @@
 
 #include <stdio.h>
 #include "limits.h"
+#include "types.h"
 
 typedef struct {
 	int n_bins;
 	int print_label;
 	char dimension_label[20];
-	double *bin_edges;
+	float64_t *bin_edges;
 } edges_t;
 
 typedef struct {
@@ -22,7 +23,7 @@ typedef struct {
 
 edges_t *allocate_edges(int n_bins);
 void free_edges(edges_t **edges);
-int edges_get_index(edges_t *edges, long long value);
+int edges_get_index(edges_t *edges, int64_t value);
 int edges_from_limits(edges_t *edges, limits_t *limits, int scale);
 void print_edges(FILE *out_stream, edges_t *edges);
 
@@ -32,11 +33,11 @@ void free_gn_histogram(gn_histogram_t **histogram);
 int gn_histogram_make_n_bins(gn_histogram_t *histogram);
 int gn_histogram_make_index_bases(gn_histogram_t *histogram);
 int gn_histogram_get_index(gn_histogram_t *histogram, 
-		long long *values);
+		int64_t *values);
 int gn_histogram_get_index_from_indices(gn_histogram_t *histogram,
 		int *indices);
 int gn_histogram_increment(gn_histogram_t *histogram,
-		long long *values);
+		int64_t *values);
 void print_gn_histogram(FILE *out_stream, gn_histogram_t *histogram);
 int gn_histogram_next_index(gn_histogram_t *histogram, int *indices);
