+++ src/intensity.h
@@ -6,7 +6,7 @@
 
 typedef struct {
 	int channels;
-	long long *counts;
+	int64_t *counts;
 } counts_t;
 
 int intensity_dispatch(FILE *in_stream, FILE *out_stream, options_t *options);
@@ -15,7 +15,7 @@ counts_t *allocate_counts(int channels);
 void init_counts(counts_t *counts);
 void free_counts(counts_t **counts);
 int increment_counts(counts_t *counts, int channel);
-void print_counts(FILE *out_stream, long long lower_time, 
-		long long upper_time, counts_t *counts);
+void print_counts(FILE *out_stream, int64_t lower_time, 
+		int64_t upper_time, counts_t *counts);
 
 #endif
